moved sha1.hpp and sha256.hpp to aux
